fixed several things, like tools.display_
and checked again the direction of the quiver in both options
only widim uses somewhat different notation, having top left corner with y.max() values
pyprocess and process both do not change y pixel values (0,0 is top left) and make
v direction according to the notation that the quiver shows left to right as positive velocity in 'x' direction, u > 0 and from top to bottom is a vertical velocity that should be "negative" and therefore in plot with 0,0 at the bottom it points upwards. to show quiver with 0,0 at the bottom, we plot "negative" v so it's upwards.
when the flow in the image rotates clockwise it means that the 0,0 top and the flow is from left to right and from top to bottom on the right side of the image (from small pixel values to the large pixel values). and the flow is from bottom to top on the left side of the image (from large pixels to small pixels)

the flow from small pixel values to large pixel values is a positive vertical velocity (y values increase). Keep it so.