Switche sign of deformation field
Previously: Deformations in x and z-direction (v,w) had a negative sign, while deformations in y-direction (u) had a positive sign.
We switched the signs in order to match the conventions in 2D PIV:  Deformations in x and z-direction have a positive sign and deformations in y-direction have a negative sign.