Update to quiver_3D
We added automatic scaling of the deformation arrows: You can use the arrow_scale parameter to scale the arrow so that the longest arrow is displayed with length of the longest axis * arrow_scale length.
We added a function to stretch the figure axis so that they all have the same scale (equal_ax).