Added new window deformation method
I added a new method for the window deformation. The new option allows to choose wether one would like to deform bot images by half of the displacement (option: 'symetric') or deform just the second image with the whole displacement( option: 'second image'). Therfore, another parameter in the settings which is called "deformation_method" was added. Default is now 'symetric'. It's only a small change.