Write output using scientific notation
The velocities were previously hard cut off if too low. This is an
unnecessary restrictions for users and leads to misleading confidence
in large velocities, as more digits are preserved. Using scientific
notation solves this problem.

Compared to the previous notation ten more characters are used per
line (that is two additional characters per value). However, the
benefit of handling a wider range of velocities outweighs this in
my opinion.

For the test to still pass, the threshold of image comparison needed
to be adjusted. This is due to the higher round-off error with the
old notation used in the reference image (specifically in 'u').